+++ webview-ui/src/components/chat/OptionsButtons.tsx
@@ -1,0 +1,69 @@
+import styled from "styled-components"
+import { CODE_BLOCK_BG_COLOR } from "../common/CodeBlock"
+import { vscode } from "../../utils/vscode"
+
+const OptionButton = styled.button<{ isSelected?: boolean; isNotSelectable?: boolean }>`
+	padding: 8px 12px;
+	background: ${(props) => (props.isSelected ? "var(--vscode-focusBorder)" : CODE_BLOCK_BG_COLOR)};
+	color: ${(props) => (props.isSelected ? "white" : "var(--vscode-input-foreground)")};
+	border: 1px solid var(--vscode-editorGroup-border);
+	border-radius: 2px;
+	cursor: ${(props) => (props.isNotSelectable ? "default" : "pointer")};
+	text-align: left;
+	font-size: 12px;
+
+	${(props) =>
+		!props.isNotSelectable &&
+		`
+		&:hover {
+			background: var(--vscode-focusBorder);
+			color: white;
+		}
+	`}
+`
+
+export const OptionsButtons = ({
+	options,
+	selected,
+	isActive,
+}: {
+	options?: string[]
+	selected?: string
+	isActive?: boolean
+}) => {
+	if (!options?.length) return null
+
+	const hasSelected = selected !== undefined && options.includes(selected)
+
+	return (
+		<div
+			style={{
+				display: "flex",
+				flexDirection: "column",
+				gap: "8px",
+				paddingTop: 15,
+				// marginTop: "22px",
+			}}>
+			{/* <div style={{ color: "var(--vscode-descriptionForeground)", fontSize: "11px", textTransform: "uppercase" }}>
+				SELECT ONE:
+			</div> */}
+			{options.map((option, index) => (
+				<OptionButton
+					key={index}
+					isSelected={option === selected}
+					isNotSelectable={hasSelected || !isActive}
+					onClick={() => {
+						if (hasSelected || !isActive) {
+							return
+						}
+						vscode.postMessage({
+							type: "optionsResponse",
+							text: option,
+						})
+					}}>
+					{option}
+				</OptionButton>
+			))}
+		</div>
+	)
+}